include two tests for orderfilter
